arranged primary color
