+++ src/navigation/Home/ExamDetail.js
@@ -1,6 +1,6 @@
-import React, {Fragment} from 'react';
+import React, { Fragment } from 'react';
 
-import {MyStack} from '../AppNavigation';
+import { MyStack } from '../AppNavigation';
 
 import ExamDetail from '@apexapp/screens/Pages/ExamDetail';
 
@@ -8,6 +8,7 @@ const ExamDetailRouter = () => {
   return (
     <Fragment>
       <MyStack.Screen name="ExamDetail" component={ExamDetail} />
+
     </Fragment>
   );
 };
