+++ src/navigation/Profile/Profile.js
@@ -1,16 +1,16 @@
-import React, {Fragment} from 'react';
+import React, { Fragment } from 'react';
 
-import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import {NavigationContainer} from '@react-navigation/native';
-import {MyStack} from '../AppNavigation';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { NavigationContainer } from '@react-navigation/native';
+import { MyStack } from '../AppNavigation';
 
 import Profile from '@apexapp/screens/Profile/Profile';
 
 const ProfileRouter = () => {
   return (
-    <Fragment>
+    <MyStack.Navigator screenOptions={{ headerShown: false }}>
       <MyStack.Screen name="Profile" component={Profile} />
-    </Fragment>
+    </MyStack.Navigator>
   );
 };
 export default ProfileRouter;
