+++ src/styles/_library.scss
@@ -7,7 +7,7 @@ $tilte1-bcr: #EAEAEA;
 $input-color: white;
 
 //colors
-$primary: #000000;
+$primary: #252775;
 $white : #ffffff;
 $black : #000000;
 $input : #F8F8F8;
@@ -20,8 +20,8 @@ $disabled: #909090;
 $h1: 35;
 $h2: 24;
 $h3: 20;
-$h4: 18;
 
 $p: 16;
-$p1: 12;
+$p1: 13;
+
 $carousel-background: #EAEAEA;
